# packages/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Ramon Inteligência Financeira — API",
  "description": "Consolidar, analisar, simular, alertar e explicar. A IA interpreta; os motores calculam.",
  "version": "1.0.0"
 },
 "paths": {
...
  "/v1/admin/crm/customers": {
   "get": {
    "summary": "Crm Customers",
    "operationId": "crm_customers_v1_admin_crm_customers_get",
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Q"
      }
     },
     {
      "name": "stage",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Stage"
      }
     },
     {
      "name": "plan",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Plan"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/admin/crm/customers/{customer_id}": {
   "get": {
    "summary": "Crm Customer",
    "operationId": "crm_customer_v1_admin_crm_customers__customer_id__get",
    "parameters": [
     {
      "name": "customer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Customer Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "summary": "Crm Update",
    "operationId": "crm_update_v1_admin_crm_customers__customer_id__patch",
    "parameters": [
     {
      "name": "customer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Customer Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CrmUpdateIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/admin/crm/customers/{customer_id}/notes": {
   "post": {
    "summary": "Crm Note",
    "operationId": "crm_note_v1_admin_crm_customers__customer_id__notes_post",
    "parameters": [
     {
      "name": "customer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Customer Id"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CrmNoteIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/admin/crm/customers/{customer_id}/payments": {
   "post": {
    "summary": "Crm Payment",
    "operationId": "crm_payment_v1_admin_crm_customers__customer_id__payments_post",
    "parameters": [
     {
      "name": "customer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Customer Id"
      }
     },
     {
      "name": "idempotency-key",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Idempotency-Key"
      }
     },
     {
      "name": "authorization",
      "in": "header",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Authorization"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CrmPaymentIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "CrmNoteIn": {
    "properties": {
     "text": {
      "type": "string",
      "maxLength": 2000,
      "minLength": 1,
      "title": "Text"
     },
     "kind": {
      "type": "string",
      "title": "Kind",
      "default": "nota"
     }
    },
    "type": "object",
    "required": [
     "text"
    ],
    "title": "CrmNoteIn"
   },
   "CrmPaymentIn": {
    "properties": {
     "amount": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "string",
        "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
       }
      ],
      "title": "Amount"
     },
     "method": {
      "type": "string",
      "title": "Method",
      "default": "pix"
     },
     "status": {
      "type": "string",
      "title": "Status",
      "default": "pago"
     },
     "date": {
      "type": "string",
      "format": "date",
      "title": "Date"
     },
     "period": {
      "type": "string",
      "title": "Period"
     },
     "reference": {
      "type": "string",
      "title": "Reference",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "amount",
     "date",
     "period"
    ],
    "title": "CrmPaymentIn"
   },
   "CrmUpdateIn": {
    "properties": {
     "stage": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stage"
     },
     "clear_override": {
      "type": "boolean",
      "title": "Clear Override",
      "default": false
     },
     "tags": {
      "anyOf": [
       {
        "items": {
         "type": "string"
        },
        "type": "array"
       },
       {
        "type": "null"
       }
      ],
      "title": "Tags"
     },
     "next_action": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Action"
     },
     "next_action_date": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Next Action Date"
     },
     "plan": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plan"
     }
    },
    "type": "object",
    "title": "CrmUpdateIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}